FPMOSC.ENH: Create package for filling pattern measurement separated from CurrInfo.
