+++ siriuspy/siriuspy/fpmosc/__init__.py
@@ -1,0 +1,3 @@
+"""CurrentInfo subpackage."""
+
+from .main import FPMOscApp
